refactor: extract StreamStorageRegistry
